refactor: extract megachurch helpers composable
