+++ src/views/megachurch/ts/_chatMessages.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 export interface ChatMessages {
   plug: {
     casual: {
